requirements.txt
#libraries
streamlit
google-generativeai
python-dotenv